requirements.txt
python-dotenv>=1.0.0
apache-airflow==2.6.0
pandas==2.2.2
requests==2.32.3
awswrangler==3.9.1
redshift-connector==2.1.3